# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Photogood Open API Docs",
        "description": "Photogood Open API",
        "contact": {
            "email": "laksamana.arya1412@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Photogood API Server"
        }
    ],
    "paths": {
        "/get-all-user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get All User API's",
                "operationId": "c17b79cfeb99bc8b529d49b63964c409",
                "responses": {
                    "200": {
                        "description": "Successfully Login"
                    },
                    "201": {
                        "description": "Successfully Login"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/get-all-member": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get All Member API's",
                "operationId": "bd78ab6185f23d5a7a7ec2e365180f86",
                "responses": {
                    "200": {
                        "description": "Successfully Login"
                    },
                    "201": {
                        "description": "Successfully Login"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login API's",
                "operationId": "c8dfeded84a9684f296d87247d968f02",
                "requestBody": {
                    "description": "- Login to your account",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Username": {
                                        "type": "string",
                                        "example": "ogud"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Login"
                    },
                    "201": {
                        "description": "Successfully Login"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Register API's",
                "operationId": "60d7f88ad035d82b4cf1b070d2b11d77",
                "requestBody": {
                    "description": "- Register to your account",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Username": {
                                        "type": "string",
                                        "example": "ogud"
                                    },
                                    "NamaLengkap": {
                                        "type": "string",
                                        "example": "ogud laksamana"
                                    },
                                    "Password": {
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "Email": {
                                        "type": "string",
                                        "example": "laksamana.arya1412@gmail.com"
                                    },
                                    "Alamat": {
                                        "type": "string",
                                        "example": "Jl sama kamu kapan?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Register"
                    },
                    "201": {
                        "description": "Successfully Register"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    }
}